# Tweak satisfiability check of the remainder branch