This PR makes sure the remainder predicate is checked considering the available side condition.