@geo2a 2024-06-20
A KEVM performance run with 1 worker isn't showing any difference:
```
| Test                                           | georgy-check-remainder-consider-side-condition time | master-974ab83b0 time | (georgy-check-remainder-consider-side-condition/master-974ab83b0) time
|------------------------------------------------|-----------------------------------------------------|-----------------------|------------------------------------------------------------------------
| benchmarks/ecrecoverloop02-sig0-invalid-spec.k | 210.77                                              | 221.67                | 0.9508278071006452
| mcd/dsvalue-peek-pass-rough-spec.k             | 50.85                                               | 48.99                 | 1.0379669320269442
| mcd/flopper-tick-pass-rough-spec.k             | 234.06                                              | 223.94                | 1.0451906760739484
| mcd/dstoken-approve-fail-rough-spec.k          | 89.51                                               | 85.45                 | 1.0475131655939145
| benchmarks/ecrecoverloop02-sigs-valid-spec.k   | 431.36                                              | 407.54                | 1.0584482504784807
| TOTAL                                          | 1016.5500000000001                                  | 987.5899999999999     | 1.0293239097196207
```