@geo2a 2024-06-20
Closing this for now, as it does not bring any benefit, and we are likely to tweak this behavior in a slightly different way in other PRs.